Fix for issue #5
This is caused by a possible bug when Dynlink links object code.  The
main problem is that camlCamlinternalOO is not linked when the
olympic.ml code is compiled.  To force linking, I have added a line
that will cause a warning to be emitted and a comment explaining the
problem.