Exprimental support for outcome lists
The list of outcomes is stored in an abstract module.  This disallows
the function that the user provides to the library from changing the
past.  However, this does allow the function to wipe the history
entirely.  This is unlikely because it would negate the usefulness of
the library as a whole.